fix(test): set organization once at the topmost level in tests to avoid Dynflow errors
Setting the current organization in tests caused a mismatch between the
organization used to start a Dynflow action and the one set when the action
completed. This led to errors like:

"/parallel-executor-core: Couldn't find Organization with 'id'=<447626448> [WHERE "taxonomies"."type" = $1] (ActiveRecord::RecordNotFound)"

This change fixes the issue by setting the organization once in the
ForemanAnsibleDirectorTestCase setup, ensuring consistency throughout the test lifecycle.